Extract mediator
